+++ README.md
@@ -1,2 +1,4 @@
-# impossiblebigoquiz
-The impossible big o quiz
+# constcodeprep
+Const Code Prep
+
+This repository is not complete.  An essential file has been added to the gitignore.  It holds the tranporter for the mailer, and the transporter contains sensitive data, password, that I do not want on github.
